i#06 First functional basic web version

+++ MyVet.Web/Controllers/PetTypesController.cs
@@ -125,23 +125,22 @@ namespace MyVet.Web.Controllers
             }
 
             var petType = await _context.PetTypes
+                .Include(pt => pt.Pets)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (petType == null)
             {
                 return NotFound();
             }
 
-            return View(petType);
-        }
+            if (petType.Pets.Count > 0)
+            {
+                ModelState.AddModelError(string.Empty, "The Pet Type can not be deleted becasuse it has related records.");
+                return RedirectToAction(nameof(Index));
+            }
 
-        // POST: PetTypes/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var petType = await _context.PetTypes.FindAsync(id);
             _context.PetTypes.Remove(petType);
             await _context.SaveChangesAsync();
+
             return RedirectToAction(nameof(Index));
         }
